Put -lQt5PlatformSupport behind a conditional and update build doc
Linking with `-lQt5PlatformSupport` is necessary inside the Docker
container which has Qt 5.6. Unfortunately that breaks the build on
Kendall's computer, which has Qt 5.8. Now the library is only linked
if Qt's version is less than 5.7.

Also updated the README to recommend building in a seperate directory
so that build files don't clutter up the source directory. Cleaned up
the .gitignore based on that recommendation.

This seems like a good place to note that I had wanted to convert the
build system from qmake to cmake, but it ended up being more
complicated than I wanted due to bugs in the mingw CMake files. Lots
of link problems and issues with platform plugins. I'm going to close
OVER-4906 and stick with qmake for now.