Add "plog" logging library
plog is a pretty simple header-only logging library. This commit adds
it as a submodule (pointed at the latest release tag), and initializes
the library prior to starting the application. This commit does not
convert existing log messages from `qDebug()` / `printf` in order to
avoid potential merge conflicts.

It's configured to log to both the application data directory and to
the console (if present). The library has an output for the Windows
event log, we can consider enabling that at some point in the future.

To use it in gondar, include `"log.h"` and make use of one of the
macros such as `LOG_DEBUG` or `LOG_INFO`. Full documentation here:

https://github.com/SergiusTheBest/plog