Pass file path to the unzipper instead of a URL
This is needed because the download manager sets a specific directory
for the downloaded files rather than using the current working
directory. The unzipper should not try to guess the path of the
downloaded file from the URL; we can just tell it directly.

Note that this does not fully fix path issues with the unzipper; more
changes in `neverware_unzipper.cc` are needed. This is a necessary
first step though.