Makefile: improve configuration variables
The configuration variables all use `?=` assignment now, which means
they can be set via environment variables. The default value will only
be assigned if the variable is not set in the environment.

The cmake program can now be set via the `CMAKE` variable.

The clang-format program can now be set via the `CLANG_FORMAT`
variable.

If `CLANG_FORMAT` is not set, and if `clang-format` does not exist,
`clang-format-4.0` is used, which will hopefully work on Kendall's
installation of Debian Untestable.