Fix scary misuse of sprintf
I'm not sure how this happened, but the arguments being passed to
sprintf are absolutely not valid. It'd try to write into a string
literal. Upstream rufus has `suprintf` for these calls, not
`sprintf`. `suprintf` appears to be part of their logging
utilities. I'm guessing there was just a find-replace error in the
conversion to gondar.

For now I've replaced these with `printf`; at some point we should
probably make it go to the gondar log.