Refactor compiler flag configuration
Put compiler warnings in a new property to make sharing easier.

This change ensures that minizip doesn't use the extra warnings,
except for minizip_extra which explicitly opts out of some.

The `-Wsuggest-override` warning is C++ only, so it's gated with a
cmake generator expression.

The mingw stdio functions are now defined project-wide rather than
just in the win32_usb target.

The win32_usb target now uses all the same warnings as the app
library.