Change default metrics api key to an empty string
Turns out that the CMake `set` command in `CACHE` mode allows you to
leave out the default value, in which case the variable doesn't get
set automatically. If not set, the `configure_file` step doesn't
`#define` the variable, and instead leaves a comment in the generated
header:

    /* #undef METRICS_API_KEY */

This means we don't need to hardcode a string like "notset" in
multiple places, instead we can just check if METRICS_API_KEY is
defined or not.

Added a `getMetricsApiKey` function to handle this.