Move some GondarWizard members out of the public header
The `gondarwizard.h` header currently includes most of the other
headers in gondar, and is itself included in many places. This means
that a lot of recompilation is required for many trivial changes, and
compilation units are bigger than they need to be. Memory usage gets
pretty high when compiling with optimization enabled, which was
causing OOM failures on the Jenkins builder. We increased the size of
the builder, but hopefully we can also make compilation less memory
hungry by reducing the size of compilation units.

This commit adds a `Private` nested class to `GondarWizard` to contain
members that don't need to be accessed outside of `GondarWizard`. As
we improve the shared-state API we can move the rest of the pages into
that `Private` class, which should leave the `gondarwizard.h` header
much lighter weight.